Move the additional methods to the helper and modify it
