+++ app/controllers/transactions_controller.rb
@@ -1,4 +1,7 @@
+require_relative "../helpers/transactions_helper"
+
 class TransactionsController < ApplicationController
+  include TransactionsHelper
   before_action :set_transaction, only: %i[ show edit update destroy ]
 
   # GET /transactions or /transactions.json
@@ -21,39 +24,6 @@ class TransactionsController < ApplicationController
   # GET /transactions/1/edit
   def edit
     @portfolio = Portfolio.find(params[:portfolio_id])
-  end
-
-  def is_enough_cash?(transaction)
-    @cash_position = Position.where(portfolio_id: params[:portfolio_id], symbol: "Cash").first
-    transaction.commission == nil ? transaction.commission = 0 : transaction.commission
-    transaction.fee == nil ? transaction.fee = 0 : transaction.fee
-    @transaction_cost = transaction.quantity * transaction.price + transaction.commission + transaction.fee
-    @cash_position.quantity >= @transaction_cost
-  end
-
-  def create_new_position
-    @portfolio = Portfolio.find(params[:portfolio_id])
-    @positions = Position.where(portfolio_id: params[:portfolio_id])
-    # @cash_position = Position.where(portfolio_id: params[:portfolio_id], symbol: "Cash").first
-
-    case @transaction.tr_type
-    when "Buy"
-      if is_enough_cash?(@transaction)
-        @positions.each do |position|
-          position_total = position.quantity * position.cost_per_share
-          if position.symbol == @transaction.symbol
-            position.update(quantity: position.quantity + @transaction.quantity)
-            position.update(cost_per_share: (position_total + @transaction_cost) / position.quantity)
-            @cash_position.update(quantity: @cash_position.quantity - (@transaction.quantity * @transaction.price + @transaction.commission + @transaction.fee))
-            break
-          else
-            # new_position = Position.new(open_date: @transaction.trade_date, symbol: @transaction.symbol, quantity: @transaction.quantity, cost_per_share: @transaction.price, portfolio_id: @portfolio.id)
-            # new_position.save
-            # break
-          end
-        end
-      end
-    end
   end
 
   # POST /transactions or /transactions.json
